backend/requirements.txt
# --- Core ---
python-dotenv
asyncio

# --- Web framework ---
fastapi
uvicorn[standard]
httpx

# --- Database / Supabase ---
supabase
postgrest
gotrue
storage3

# --- LLM / AI integrations ---
anthropic
google-generativeai
livekit-agents
livekit-plugins-google
livekit-plugins-deepgram
livekit-plugins-anthropic
livekit-plugins-noise-cancellation

# --- JWT / Security ---
PyJWT

# --- Data / Utils ---
pydantic
typing-extensions
python-dateutil
base58

# --- Image handling ---
Pillow

# --- Optional (helpful during dev) ---
requests
